`aset-*` (`aset-int`, `aset-long`, etc.) is a set of reflective array operations that are tens of times slower than inline `aset` calls and can have a significant performance impact:

```clojure
user=> (let [arr (int-array 1000)] (cr/quick-bench (dotimes [i 1000] (aset arr i i))))
Evaluation count : 587442 in 6 samples of 97907 calls.
             Execution time mean : 1.012114 µs
    Execution time std-deviation : 5.606968 ns
   Execution time lower quantile : 1.005990 µs ( 2.5%)
   Execution time upper quantile : 1.020545 µs (97.5%)
                   Overhead used : 6.755950 ns
nil
user=> (let [arr (int-array 1000)] (cr/quick-bench (dotimes [i 1000] (aset-int arr i i))))
Evaluation count : 8418 in 6 samples of 1403 calls.
             Execution time mean : 73.125862 µs
    Execution time std-deviation : 1.971431 µs
   Execution time lower quantile : 71.480682 µs ( 2.5%)
   Execution time upper quantile : 75.451133 µs (97.5%)
                   Overhead used : 6.755950 ns
nil
user=>
```

This PR replaces all the invocations to `aset-*` in the codebase with `aset` for performance improvement.

This change contributes specifically to the performance improvement of `cljam.algo.depth`, which is more than 1.5x faster after the change:

```clojure
;; before change
user=> (time (dorun (with-open [r (sam/reader "sample.bam")] (depth/depth r {:chr "chr1"} {:unchecked? true}))))
"Elapsed time: 41775.877042 msecs"
nil
user=>

;; after change
user=> (time (dorun (with-open [r (sam/reader "sample.bam")] (depth/depth r {:chr "chr1"} {:unchecked? true}))))
"Elapsed time: 24674.470916 msecs"
nil
user=>
```

| before change | after change |
| :---: | :---: |
| <img width="400" alt="flamegraph-before-change" src="https://github.com/chrovis/cljam/assets/27441/f546f275-bc2b-44bb-9735-e3907d0fff66"> | <img width="400" alt="flamegraph-after-change" src="https://github.com/chrovis/cljam/assets/27441/e2428139-9fd7-4a3d-b031-f534f2af5fa5"> |