@NoahTheDuke 2024-05-21
This is being tracked in [CLJ-2840](https://clojure.atlassian.net/browse/CLJ-2840), so hopefully such changes won't be necessary in the future.