@athos 2024-05-21
@NoahTheDuke  Thank you for letting us know! We are already watching the ticket and hope that fix will further improve Clojure.

However, for this library, similar fixes might still be necessary in the future since we need to support older versions of Clojure as well.